pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-skill-kit"
version = "0.1.6"
description = "Centralized skills repository and CLI toolkit for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
keywords = ["ai", "agents", "skills", "cli", "codex", "gemini", "claude"]
authors = [
    { name = "Nava" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
ask = "ask.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["ask", "agents", "skills"]
